Refactor common probe metadata class
